Perform big A/E/T rename, make B/* flat products
T for Tabulate frees up * to produce a flat product, which was
empirically the far more common use in AoC 2018, and I think is also a
tad closer in spirit to the other overloads of *. T also gives access to
potentially more useful coercion cases. Meanwhile, A for 10 is more
consistent with literal shortcuts; E for Exit is more consistent with
general computer terminology.